Moved connection service from plugin to libswganh
