+++ src/CMakeLists.txt
@@ -107,7 +107,6 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
 add_subdirectory(anh)
 
 ## project plugins
-add_subdirectory(connection)
 add_subdirectory(character)
 
 ## project executables
