--- src/connection/CMakeLists.txt
@@ -1,47 +1,0 @@
-include(ANHPlugin)
-
-AddANHPlugin(connection
-    DEPENDS 
-        libanh
-        libswganh
-    ADDITIONAL_INCLUDE_DIRS
-        ${Boost_INCLUDE_DIR}
-        ${MYSQL_INCLUDE_DIR}
-        ${MYSQLCONNECTORCPP_INCLUDE_DIRS}
-        ${GLM_INCLUDE_DIR}
-        ${GLOG_INCLUDE_DIR}
-        ${PYTHON_INCLUDE_DIR}
-        ${TBB_INCLUDE_DIRS}
-		${ZLIB_INCLUDE_DIR}
-    ADDITIONAL_LIBRARY_DIRS
-        ${Boost_LIBRARY_DIRS}
-        ${TBB_LIBRARY_DIRS}
-    DEBUG_LIBRARIES 
-        ${MYSQL_LIB}
-        ${MYSQLCONNECTORCPP_LIBRARY_DEBUG}
-        ${TBB_DEBUG_LIBRARIES}
-        ${Boost_DATE_TIME_LIBRARY_DEBUG}
-        ${Boost_PROGRAM_OPTIONS_LIBRARY_DEBUG}
-        ${Boost_PYTHON_LIBRARY_DEBUG}
-        ${Boost_REGEX_LIBRARY_DEBUG}
-        ${Boost_SYSTEM_LIBRARY_DEBUG}
-        ${Boost_THREAD_LIBRARY_DEBUG}
-        ${Boost_FILESYSTEM_LIBRARY_DEBUG}
-        ${PYTHON_LIBRARY}
-        ${GLOG_LIBRARY_DEBUG}
-		${ZLIB_LIBRARY_DEBUG}
-    OPTIMIZED_LIBRARIES 
-        ${MYSQL_LIB}
-        ${MYSQLCONNECTORCPP_LIBRARY_RELEASE}
-        ${Boost_DATE_TIME_LIBRARY_RELEASE}
-        ${Boost_PROGRAM_OPTIONS_LIBRARY_RELEASE}
-        ${Boost_PYTHON_LIBRARY_RELEASE}
-        ${Boost_REGEX_LIBRARY_RELEASE}
-        ${Boost_SYSTEM_LIBRARY_RELEASE}
-        ${Boost_THREAD_LIBRARY_RELEASE}
-        ${Boost_FILESYSTEM_LIBRARY_RELEASE}
-        ${TBB_LIBRARIES}
-        ${PYTHON_LIBRARY}
-        ${GLOG_LIBRARY_RELEASE}
-		${ZLIB_LIBRARY_RELEASE}
-)
