--- src/connection/main.cc
@@ -1,86 +1,0 @@
-/*
- This file is part of SWGANH. For more information, visit http://swganh.com
- 
- Copyright (c) 2006 - 2011 The SWG:ANH Team
-
- This program is free software; you can redistribute it and/or
- modify it under the terms of the GNU General Public License
- as published by the Free Software Foundation; either version 2
- of the License, or (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program; if not, write to the Free Software
- Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-*/
- 
-#include <iostream>
-
-#include <boost/program_options/options_description.hpp>
-#include <boost/program_options/variables_map.hpp>
-#include <boost/thread.hpp>
-
-#include <glog/logging.h>
-
-#include "anh/app/kernel_interface.h"
-#include "anh/plugin/bindings.h"
-#include "anh/plugin/plugin_manager.h"
-#include "anh/service/service_manager.h"
-
-#include "connection/connection_service.h"
-
-using namespace anh;
-using namespace app;
-using namespace event_dispatcher;
-using namespace connection;
-using namespace plugin;
-using namespace swganh::character;
-using namespace std;
-using boost::program_options::options_description;
-using boost::program_options::variables_map;
-
-static struct ConnectionConfig {
-    string listen_address;
-    uint16_t listen_port;
-} connection_config;
-
-extern "C" PLUGIN_API void ExitModule() {
-    return;
-}
-
-extern "C" PLUGIN_API ExitFunc InitializePlugin(shared_ptr<KernelInterface> kernel) {
-
-    ObjectRegistration registration;
-    registration.version.major = 1;
-    registration.version.minor = 0;
-
-    // Register TestObj
-    registration.CreateObject = [] (ObjectParams* params) -> void * {
-        auto connection_service = new ConnectionService(
-            connection_config.listen_address, connection_config.listen_port, params->kernel);
-        return connection_service;
-    };
-
-    registration.DestroyObject = [] (void * object) {
-        if (object) {
-            delete static_cast<ConnectionService*>(object);
-        }
-    };
-
-    kernel->GetPluginManager()->RegisterObject("ConnectionService", &registration);
-
-    return ExitModule;
-}
-
-extern "C" PLUGIN_API void ConfigurePlugin(options_description& description) {
-    description.add_options()
-        ("service.connection.udp_port", boost::program_options::value<uint16_t>(&connection_config.listen_port),
-            "The port the connection service will listen for incoming client connections on")
-        ("service.connection.address", boost::program_options::value<string>(&connection_config.listen_address),
-            "The public address the connection service will listen for incoming client connections on")
-    ;
-}
