+++ src/swganh/connection/connection_service.h
@@ -1,6 +1,6 @@
 
-#ifndef SWGANH_CONNECTION_BASE_CONNECTION_SERVICE_H_
-#define SWGANH_CONNECTION_BASE_CONNECTION_SERVICE_H_
+#ifndef SWGANH_CONNECTION_CONNECTION_SERVICE_H_
+#define SWGANH_CONNECTION_CONNECTION_SERVICE_H_
 
 #include <cstdint>
 #include <memory>
@@ -19,6 +19,13 @@
 
 #include "swganh/character/base_character_service.h"
 #include "swganh/login/login_service.h"
+#include "swganh/scene/messages/cmd_scene_ready.h"
+
+#include "swganh/connection/providers/session_provider_interface.h"
+
+#include "swganh/connection/messages/client_permissions_message.h"
+#include "swganh/connection/messages/client_id_msg.h"
+#include "swganh/connection/messages/heart_beat.h"
 
 namespace anh {
 namespace network {
@@ -31,7 +38,7 @@ namespace connection {
     
 struct ConnectionClient;
     
-class BaseConnectionService 
+class ConnectionService 
     : public swganh::base::BaseService
     , public swganh::base::SwgMessageRouter<ConnectionClient>
 {
@@ -42,10 +49,14 @@ public:
         anh::network::soe::EndpointHashCompare
     > ClientMap;
 public:
-    BaseConnectionService(
+    ConnectionService(
         std::string listen_address, 
         uint16_t listen_port, 
         std::shared_ptr<anh::app::KernelInterface> kernel);
+
+    anh::service::ServiceDescription GetServiceDescription();
+
+    void subscribe();
 
     void onStart();
     void onStop();
@@ -69,6 +80,14 @@ protected:
 private:    
     ClientMap clients_;
     
+    void HandleClientIdMsg_(std::shared_ptr<swganh::connection::ConnectionClient> client, const swganh::connection::messages::ClientIdMsg& message);
+    void HandleCmdSceneReady_(std::shared_ptr<swganh::connection::ConnectionClient> client, const swganh::scene::messages::CmdSceneReady& message);
+    
+    void RemoveClient_(std::shared_ptr<anh::network::soe::Session> session);
+    void AddClient_(uint64_t player_id, std::shared_ptr<swganh::connection::ConnectionClient> client);
+
+    std::shared_ptr<providers::SessionProviderInterface> session_provider_;
+    
     std::unique_ptr<anh::network::soe::Server> soe_server_;
     
     std::shared_ptr<swganh::character::BaseCharacterService> character_service_;
@@ -80,4 +99,4 @@ private:
     
 }}  // namespace swganh::connection
 
-#endif  // SWGANH_CONNECTION_BASE_CONNECTION_SERVICE_H_
+#endif  // SWGANH_CONNECTION_CONNECTION_SERVICE_H_
